backend/requirements.txt
pytesseract
python-pptx
Pillow
python-dotenv
